レッスン並び替え処理サービスクラスの実装
