+++ app/Http/Controllers/Api/Instructor/LessonController.php
@@ -26,6 +26,8 @@ use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
+use App\Services\Lesson\SortLessonsService;
+
 
 /**
  * @tags Instructor-Lesson
@@ -327,7 +329,7 @@ class LessonController extends Controller
     /**
      * レッスン並び替えAPI
      */
-    public function sort(SortRequest $request): JsonResponse
+    public function sort(SortRequest $request, SortLessonsService $sortLessonsService): JsonResponse
     {
         DB::beginTransaction();
 
@@ -357,14 +359,7 @@ class LessonController extends Controller
                 }
             });
 
-            // orderカラムを更新（並び替え実施）
-            $lessons->each(function (Lesson $lesson) use ($inputLessons) {
-                $collectionLessons = new Collection($inputLessons);
-                $inputLesson = $collectionLessons->firstWhere('lesson_id', $lesson->id);
-                $lesson->update([
-                    'order' => $inputLesson['order'],
-                ]);
-            });
+            $sortLessonsService($lessons, $inputLessons);
 
             DB::commit();
